+++ game/scrabble.py
@@ -1,6 +1,7 @@
 from game.board import Board
 from game.player import Player
-from game.tiles import BagTiles
+from game.tiles import BagTiles, Tile
+from game.cell import Cell
 from game.dictionary import validate_word as validate_word_dict
 
 class InvalidWord(Exception):
@@ -39,6 +40,59 @@ class ScrabbleGame:
             raise InvalidPlaceWordException("No es correcta la ubicación")
         if not self.board.validate_word_place_board(word, location, orientaiton):
             raise InvalidPlaceWordException("No se puede colocar")
+    
+    def get_tile_value(self, letter):
+        for tile in self.bag_tiles.tiles:
+            if tile.letter == letter:
+                return tile.value
+        return 0
+    
+    def get_coordinates_for_index(self, location, orientation, index):
+        x, y = location
+        if orientation.lower() == 'h':
+            return x, y + index
+        elif orientation.lower() == 'v':
+            return x + index, y
+        else:
+            raise ValueError("Orientación no válida")
+    
+    def get_score(self, word, location, orientation):
+        total_score = 0
+        word_multiplier = 1
+
+        for i, letter in enumerate(word.upper()):
+            row, col = self.get_coordinates_for_index(location, orientation, i)
+            tile_value = self.get_tile_value(letter)
+
+            cell = self.board.grid[row][col]
+
+            if cell.active:
+                cell = Cell(letter=Tile(letter, tile_value),
+                            multiplier_type=cell.multiplier_type,
+                            multiplier=cell.multiplier)
+                cell_value = cell.calculate_value()
+                total_score += cell_value
+
+                if cell.multiplier_type == 'word':
+                    word_multiplier *= cell.multiplier
+
+        return total_score * word_multiplier
+        
+    def play(self, word, location, orientation):
+        self.validate_word(word, location, orientation)
+        self.board.put_word(word, location, orientation)
+        total_score = self.get_score(word, location, orientation)
+
+        self.players[self.current_player].rellenar()
+        self.players[self.current_player].score += total_score
+
+        x, y = location
+        for i in range(len(word)):
+            if orientation == 'H':
+                self.board.deactivate_cell(x, y + i)
+            else:
+                self.board.deactivate_cell(x + i, y)
+        self.next_turn()
 
     
     # def play(self, word, location, orientation):