+++ tests/tests_board.py
@@ -257,28 +257,6 @@ class TestBoard(unittest.TestCase):
         with self.assertRaises(SoloVoHParaLaOrientacion):
             board.put_word(word, location, orientation)
 
-    # def test_draw_board(self):
-    #     board = Board()
-    #     expected_board = (
-    #         "     1  2  3  4  5  6  7  8  9 10 11 12 13 14 15\n"
-    #         " 1|  _  _  _  _  _  _  _  _  _  _  _  _  _  _  _  \n"
-    #         " 2|  _  _  _  _  _  _  _  _  _  _  _  _  _  _  _  \n"
-    #         " 3|  _  _  _  _  _  _  _  _  _  _  _  _  _  _  _  \n"
-    #         " 4|  _  _  _  _  _  _  _  _  _  _  _  _  _  _  _  \n"
-    #         " 5|  _  _  _  _  _  _  _  _  _  _  _  _  _  _  _  \n"
-    #         " 6|  _  _  _  _  _  _  _  _  _  _  _  _  _  _  _  \n"
-    #         " 7|  _  _  _  _  _  _  _  _  _  _  _  _  _  _  _  \n"
-    #         " 8|  _  _  _  _  _  _  _  _  _  _  _  _  _  _  _  \n"
-    #         " 9|  _  _  _  _  _  _  _  _  _  _  _  _  _  _  _  \n"
-    #         "10|  _  _  _  _  _  _  _  _  _  _  _  _  _  _  _  \n"
-    #         "11|  _  _  _  _  _  _  _  _  _  _  _  _  _  _  _  \n"
-    #         "12|  _  _  _  _  _  _  _  _  _  _  _  _  _  _  _  \n"
-    #         "13|  _  _  _  _  _  _  _  _  _  _  _  _  _  _  _  \n"
-    #         "14|  _  _  _  _  _  _  _  _  _  _  _  _  _  _  _  \n"
-    #         "15|  _  _  _  _  _  _  _  _  _  _  _  _  _  _  _  \n"
-    #     )
-    #     result_board = board.draw_board()
-    #     self.assertEqual(result_board, expected_board)
 
 if __name__ == '__main__':
     unittest.main()